created calendarComponent & moved gridholder div from editor to components
