+++ src/components/ImageCatalogue.jsx
@@ -1,8 +1,10 @@
 import TextComponent from "./TextComponent"
 
-const ImageCatalogue = ({ bool, images, handleBgSelection, guideH, guideText }) => {
+const ImageCatalogue = ({ bool, images, handleBgSelection, guideH, guideText, calendarImage }) => {
     return (
-        <>
+        <div className="gridHolder" style={{
+            backgroundImage: `url(${calendarImage})`,
+        }}>
             {bool && <TextComponent guideH={guideH} guideText={guideText} />}
             {!bool &&
                 <TextComponent guideH={guideH} guideText={guideText} />
@@ -12,7 +14,7 @@ const ImageCatalogue = ({ bool, images, handleBgSelection, guideH, guideText }) 
                     backgroundImage: `url(${hatchImg.urls.small})`
                 }}></div>
             })}
-        </>
+        </div>
     )
 }
